Clean-up build_bins.sh idempotency checks

+++ build_bins.sh
@@ -52,7 +52,7 @@ FILES="kibana3.tgz $FILES"
 
 # Grab plugins for fluentd
 for i in elasticsearch tail-multiline tail-ex record-reformer rewrite; do
-  if [[ ! -f fluent-plugin-${i}*.gem && ! -f gems/fluent-plugin-${i}*.gem ]]; then
+  if ! [[ -f gems/fluent-plugin-${i}.gem ]]; then
     gem fetch fluent-plugin-${i}
     ln -s fluent-plugin-${i}-*.gem fluent-plugin-${i}.gem || true
   fi
@@ -60,20 +60,20 @@ for i in elasticsearch tail-multiline tail-ex record-reformer rewrite; do
 done
 
 # Get the Rubygem for zookeeper
-if [[ ! -f zookeeper*.gem || ! -f gems/zookeeper*.gem ]]; then
+if ! [[ -f gems/zookeeper.gem ]]; then
   gem fetch zookeeper -v 1.4.7
   ln -s zookeeper-*.gem zookeeper.gem || true
 fi
 FILES="zookeeper*.gem $FILES"
 
 # Fetch the cirros image for testing
-if [[ ! -f cirros-0.3.0-x86_64-disk.img ]]; then
+if ! [[ -f cirros-0.3.0-x86_64-disk.img ]]; then
   $CURL -O -L https://launchpad.net/cirros/trunk/0.3.0/+download/cirros-0.3.0-x86_64-disk.img
 fi
 FILES="cirros-0.3.0-x86_64-disk.img $FILES"
 
 # Grab the Ubuntu 12.04 installer image
-if [[ ! -f ubuntu-12.04-mini.iso ]]; then
+if ! [[ -f ubuntu-12.04-mini.iso ]]; then
   # Download this ISO to get the latest kernel/X LTS stack installer
   #$CURL -o ubuntu-12.04-mini.iso http://archive.ubuntu.com/ubuntu/dists/precise-updates/main/installer-amd64/current/images/raring-netboot/mini.iso
   $CURL -o ubuntu-12.04-mini.iso http://archive.ubuntu.com/ubuntu/dists/precise/main/installer-amd64/current/images/netboot/mini.iso
@@ -81,18 +81,18 @@ fi
 FILES="ubuntu-12.04-mini.iso $FILES"
 
 # Grab the CentOS 6 PXE boot images
-if [[ ! -f centos-6-initrd.img ]]; then
+if ! [[ -f centos-6-initrd.img ]]; then
   $CURL -o centos-6-initrd.img http://mirror.net.cen.ct.gov/centos/6/os/x86_64/images/pxeboot/initrd.img
 fi
 FILES="centos-6-initrd.img $FILES"
 
-if [[ ! -f centos-6-vmlinuz ]]; then
+if ! [[ -f centos-6-vmlinuz ]]; then
   $CURL -o centos-6-vmlinuz http://mirror.net.cen.ct.gov/centos/6/os/x86_64/images/pxeboot/vmlinuz
 fi
 FILES="centos-6-vmlinuz $FILES"
 
 # Make the diamond package
-if [[ ! -f diamond*.deb ]]; then
+if ! [[ -f diamond.deb ]]; then
   git clone https://github.com/BrightcoveOS/Diamond.git
   pushd Diamond
   git checkout $VER_DIAMOND
@@ -105,12 +105,12 @@ fi
 FILES="diamond.deb $FILES"
 
 # Snag elasticsearch
-if [[ ! -f elasticsearch-0.90.3.deb ]]; then
+if ! [[ -f elasticsearch-0.90.3.deb ]]; then
   $CURL -O -L https://download.elasticsearch.org/elasticsearch/elasticsearch/elasticsearch-0.90.3.deb
 fi
 FILES="elasticsearch-0.90.3.deb $FILES"
 
-if [[ ! -f elasticsearch-plugins.tgz ]]; then
+if ! [[ -f elasticsearch-plugins.tgz ]]; then
   git clone https://github.com/mobz/elasticsearch-head.git
   cd elasticsearch-head
   git archive --output ../elasticsearch-plugins.tgz --prefix head/_site/ $VER_ESPLUGIN
@@ -120,7 +120,7 @@ fi
 FILES="elasticsearch-plugins.tgz $FILES"
 
 # Fetch pyrabbit
-if [[ ! -f python/pyrabbit-1.0.1.tar.gz ]]; then
+if ! [[ -f python/pyrabbit-1.0.1.tar.gz ]]; then
   while ! $(file python/pyrabbit-1.0.1.tar.gz | grep -q 'gzip compressed data'); do
     (cd python && $CURL -O -L http://pypi.python.org/packages/source/p/pyrabbit/pyrabbit-1.0.1.tar.gz)
   done
@@ -128,9 +128,9 @@ fi
 FILES="pyrabbit-1.0.1.tar.gz $FILES"
 
 # Build graphite packages
-if [[ ! -f python-carbon_0.9.10_all.deb || \
-      ! -f python-whisper_0.9.10_all.deb || \
-      ! -f python-graphite-web_0.9.10_all.deb ]]; then
+if ! [[ -f python-carbon_0.9.10_all.deb && \
+        -f python-whisper_0.9.10_all.deb && \
+        -f python-graphite-web_0.9.10_all.deb ]]; then
   while ! $(file carbon-0.9.10.tar.gz | grep -q 'gzip compressed data'); do
     $CURL -L -O http://pypi.python.org/packages/source/c/carbon/carbon-0.9.10.tar.gz
   done
@@ -150,13 +150,13 @@ fi
 FILES="python-carbon_0.9.10_all.deb python-whisper_0.9.10_all.deb python-graphite-web_0.9.10_all.deb $FILES"
 
 # Download Python requests-aws for Zabbix monitoring
-if [[ ! -f python-requests-aws_0.1.5_all.deb ]]; then
+if ! [[ -f python-requests-aws_0.1.5_all.deb ]]; then
   fpm -s python -t deb -v 0.1.5 requests-aws
 fi
 FILES="python-requests-aws_0.1.5_all.deb $FILES"
 
 # Build the zabbix packages
-if [[ ! -f zabbix-agent.tar.gz ]] || [[ ! -f zabbix-server.tar.gz ]]; then
+if ! [[ -f zabbix-agent.tar.gz && -f zabbix-server.tar.gz ]]; then
   $CURL -L -O http://sourceforge.net/projects/zabbix/files/ZABBIX%20Latest%20Stable/2.0.7/zabbix-2.0.7.tar.gz
   tar zxf zabbix-2.0.7.tar.gz
   rm -rf /tmp/zabbix-install && mkdir -p /tmp/zabbix-install
@@ -182,7 +182,7 @@ FILES="zabbix-agent.tar.gz zabbix-server.tar.gz $FILES"
 opscode_urls="https://opscode-omnibus-packages.s3.amazonaws.com/ubuntu/12.04/x86_64/chef_11.8.0-1.ubuntu.12.04_amd64.deb
 https://opscode-omnibus-packages.s3.amazonaws.com/ubuntu/12.04/x86_64/chef-server_11.0.8-1.ubuntu.12.04_amd64.deb"
 for url in $opscode_urls; do
-  if [ ! -f $(basename $url) ]; then
+  if ! [[ -f $(basename $url) ]]; then
     $CURL -L -O $url
   fi
 done
@@ -195,7 +195,7 @@ apt-ftparchive release . > Release
 rm -f Release.gpg
 
 # generate a key and sign repo
-if [[ ! -f ${HOME}/apt_key.sec || ! -f apt_key.pub ]]; then
+if ! [[ -f ${HOME}/apt_key.sec && -f apt_key.pub ]]; then
   rm -rf ${HOME}/apt_key.sec apt_key.pub
   gpg --batch --gen-key << EOF
     Key-Type: DSA
@@ -220,20 +220,25 @@ gpg -abs --keyring ./apt_key.pub --secret-keyring ${HOME}/apt_key.sec -o Release
 # can then follow http://askubuntu.com/questions/399446
 # but can't upgrade setuptools first as:
 # "/usr/bin/pip install: error: no such option: --no-use-wheel"
-/usr/bin/pip install pip2pi || /bin/true
-/usr/local/bin/pip install setuptools --no-use-wheel --upgrade
-/usr/local/bin/pip install pip2pi 
+if ! hash dir2pi; then
+  /usr/bin/pip install pip2pi || /bin/true
+  /usr/local/bin/pip install setuptools --no-use-wheel --upgrade
+  /usr/local/bin/pip install pip2pi 
+fi
+
 dir2pi python
 
 #########################
 # generate rubygems repos
 
 # need the builder gem to generate a gem index
-gem install builder --no-ri --no-rdoc
+if [[ -z `gem list --local builder | grep builder | cut -f1 -d" "` ]]; then
+  gem install builder --no-ri --no-rdoc
+fi
 
 # place all gems into the server normally
 [ ! -d gems ] && mkdir gems
-[ -n "$(echo *.gem)" ] && mv *.gem gems
+[ "$(echo *.gem)" != '*.gem' ] && mv *.gem gems
 gem generate_index --legacy
 
 popd
